Remove dead codes
