+++ lib/content-services/content-metadata/components/content-metadata-card/content-metadata-card.component.spec.ts
@@ -26,7 +26,6 @@ import { MatExpansionModule, MatCardModule, MatButtonModule, MatIconModule } fro
 import { ContentMetadataService } from '../../services/content-metadata.service';
 import { BasicPropertiesService } from '../../services/basic-properties.service';
 import { PropertyGroupTranslatorService } from '../../services/property-groups-translator.service';
-import { PropertyDescriptorsLoaderService } from '../../services/property-descriptors-loader.service';
 import { PropertyDescriptorsService } from '../../services/property-descriptors.service';
 import { AlfrescoApiService } from '@alfresco/adf-core';
 import { ContentMetadataConfigFactory } from '../../services/config/content-metadata-config.factory';
@@ -54,7 +53,6 @@ describe('ContentMetadataCardComponent', () => {
                 ContentMetadataService,
                 BasicPropertiesService,
                 PropertyGroupTranslatorService,
-                PropertyDescriptorsLoaderService,
                 ContentMetadataConfigFactory,
                 PropertyDescriptorsService,
                 AlfrescoApiService
