+++ lib/content-services/content-metadata/content-metadata.module.ts
@@ -24,7 +24,6 @@ import { CardViewModule , FileSizePipe } from '@alfresco/adf-core';
 import { ContentMetadataComponent } from './components/content-metadata/content-metadata.component';
 import { ContentMetadataCardComponent } from './components/content-metadata-card/content-metadata-card.component';
 import { PropertyDescriptorsService } from './services/property-descriptors.service';
-import { PropertyDescriptorsLoaderService } from './services/property-descriptors-loader.service';
 import { ContentMetadataConfigFactory } from './services/config/content-metadata-config.factory';
 import { BasicPropertiesService } from './services/basic-properties.service';
 import { ContentMetadataService } from './services/content-metadata.service';
@@ -49,7 +48,6 @@ import { PropertyGroupTranslatorService } from './services/property-groups-trans
     providers: [
         ContentMetadataService,
         PropertyDescriptorsService,
-        PropertyDescriptorsLoaderService,
         ContentMetadataConfigFactory,
         BasicPropertiesService,
         PropertyGroupTranslatorService,
