+++ lib/content-services/content-metadata/services/config/layout-oriented-config.service.spec.ts
@@ -80,74 +80,6 @@ describe('LayoutOrientedConfigService', () => {
         });
     });
 
-    describe('isPropertyAllowed', () => {
-
-        const testCases = [
-            {
-                config: [],
-                expectation: false,
-                groupNameToQuery: 'berseria',
-                propertyNameToQuery: 'velvet'
-            },
-            {
-                config: [{ title: 'Deamons', items: [{ aspect: 'berseria', properties: '*' }] }],
-                expectation: true,
-                groupNameToQuery: 'berseria',
-                propertyNameToQuery: 'velvet'
-            },
-            {
-                config: [{ title: 'Deamons', items: [{ type: 'berseria', properties: ['velvet'] }] }],
-                expectation: true,
-                groupNameToQuery: 'berseria',
-                propertyNameToQuery: 'velvet'
-            },
-            {
-                config: [{ title: 'Deamons', items: [{ aspect: 'berseria', properties: ['magilou'] }] }],
-                expectation: false,
-                groupNameToQuery: 'berseria',
-                propertyNameToQuery: 'velvet'
-            },
-            {
-                config: [{ title: 'Deamons', items: [
-                    { aspect: 'berseria', properties: 'magilou' },
-                    { aspect: 'berseria', properties: 'laphicet' },
-                    { aspect: 'berseria', properties: 'velvet' }
-                ]}],
-                expectation: true,
-                groupNameToQuery: 'berseria',
-                propertyNameToQuery: 'velvet'
-            },
-            {
-                config: [
-                    { title: 'Deamons', items: [{ aspect: 'berseria', properties: 'rokurou' }] },
-                    { title: 'Malakhims', items: [{ aspect: 'berseria', properties: 'velvet' }] }
-                ],
-                expectation: true,
-                groupNameToQuery: 'berseria',
-                propertyNameToQuery: 'velvet'
-            },
-            {
-                config: [
-                    { title: 'Deamons', items: [{ aspect: 'berseria', properties: 'laphicet' }] },
-                    { title: 'Malakhims', items: [{ type: 'berseria', properties: 'rokurou' }] }
-                ],
-                expectation: false,
-                groupNameToQuery: 'berseria',
-                propertyNameToQuery: 'velvet'
-            }
-        ];
-
-        testCases.forEach((testCase, index) => {
-            it(`should return ${testCase.expectation.toString()} for test case index #${index}`, () => {
-                configService = createConfigService(testCase.config);
-
-                const isAllowed = configService.isPropertyAllowed(testCase.groupNameToQuery, testCase.propertyNameToQuery);
-
-                expect(isAllowed).toBe(testCase.expectation);
-            });
-        });
-    });
-
     describe('reorganiseByConfig', () => {
 
         interface TestCase {
